modularized server
