+++ server/.env
@@ -1,5 +1,6 @@
 TEAM_NAME="Ethan, Isabel and William"
 FLASK_ENV="development"
+SECRET_KEY="bubblegum"
 
 DB_USERNAME=""
 DB_PASSWORD=""
